refactor!: Use ES6 modules for imports
Requires web server? npx live-server
