+++ index.html
@@ -22,7 +22,6 @@
 
   <body>
     <div id="title">Mars Global Crater Database</div>
-    <!-- Dedicated title section -->
 
     <div id="filter-container">
       <div class="filter-group">
@@ -81,7 +80,6 @@
 
     <div id="map-container">
       <div id="map"></div>
-      <!-- Map container -->
     </div>
 
     <div id="table-container">
@@ -96,9 +94,9 @@
     <!-- <div id="chart">Charts</div> -->
 
     <!-- These should be loaded last -->
-    <script src="scripts/utils.js"></script>
-    <script src="scripts/slider.js"></script>
-    <script src="scripts/filter.js"></script>
-    <script src="scripts/map.js"></script>
+    <!-- <script src="scripts/utils.js"></script> -->
+    <!-- <script src="scripts/slider.js"></script> -->
+    <!-- <script src="scripts/filter.js"></script> -->
+    <script type="module" src="scripts/main.js"></script>
   </body>
 </html>
